Added Scaling for Forestation tool
so much lag tho

+++ Assets/Scripts/Data/TerrainData.cs
@@ -5,7 +5,7 @@ using UnityEngine;
     [CreateAssetMenu()]
 public class TerrainData : UpdatableData
 {
-    public float uniformScale = 2f;
+    public int uniformScale = 2;
     public float meshHeightMultiplier;
     public bool useFalloff;
